ISSUE-24[refactor] convertTimeUtil 분리
